Extract encrypted cookie tests into their own file and add more cases
